Make the snackbar global
