+++ src/components/ToastProvider/ToastProvider.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import { ToastProviderProps } from '@/types/toastProvider';
+import { ToastContainer } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+
+export default function ToastProvider({ children }: ToastProviderProps) {
+  return (
+    <>
+      {children}
+      <ToastContainer />
+    </>
+  );
+}
