+++ src/types/toastProvider.ts
@@ -1,0 +1,3 @@
+export interface ToastProviderProps {
+  children: React.ReactNode;
+}
